API 2/n: Hook up v2 routes, plus the refactoring to make this possible
A lot of work in this diff, but the high-level goal is to introduce a
system where the type system protects accidental mispairings between all
of the moving pieces that need to be accounted for with version bumps.

Some of the changes here:
- API is now a container for API version types
- Each API version defines the routes it responds to and the APIs it expects the backing record store to implement
- API versions tranform their arguments and results into data that consumers of that API expect, regardless of how the backing data changes
- The GetRecords/PostRecords functions have been generalized to take an API interface, which allows us to use multiple API version impls

Tests need to be cleaned up to account for these changes.